# openapi-spec.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/employees": {
      "post": {
        "operationId": "EmployeesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEmployeeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Funcionário criado com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Employee"
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos."
          },
          "409": {
            "description": "CPF ou email já existem."
          }
        },
        "summary": "Criar novo funcionário",
        "tags": [
          "Funcionários"
        ]
      },
      "get": {
        "operationId": "EmployeesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Lista de funcionários retornada com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Employee"
                  }
                }
              }
            }
          }
        },
        "summary": "Listar todos os funcionários",
        "tags": [
          "Funcionários"
        ]
      }
    },
    "/api/employees/{id}": {
      "get": {
        "operationId": "EmployeesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID do funcionário",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Funcionário encontrado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Employee"
                }
              }
            }
          },
          "404": {
            "description": "Funcionário não encontrado."
          }
        },
        "summary": "Buscar funcionário por ID",
        "tags": [
          "Funcionários"
        ]
      },
      "patch": {
        "operationId": "EmployeesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID do funcionário",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEmployeeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Funcionário atualizado com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Employee"
                }
              }
            }
          },
          "404": {
            "description": "Funcionário não encontrado."
          },
          "409": {
            "description": "Email já existe."
          }
        },
        "summary": "Atualizar funcionário",
        "tags": [
          "Funcionários"
        ]
      },
      "delete": {
        "operationId": "EmployeesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID do funcionário",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Funcionário removido com sucesso."
          },
          "404": {
            "description": "Funcionário não encontrado."
          }
        },
        "summary": "Remover funcionário",
        "tags": [
          "Funcionários"
        ]
      }
    },
...
    "/api/employees/positions/by-department/{departmentId}": {
      "get": {
        "operationId": "EmployeesController_getPositionsByDepartment",
        "parameters": [
          {
            "name": "departmentId",
            "required": true,
            "in": "path",
            "description": "ID do departamento",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de cargos do departamento."
          }
        },
        "summary": "Listar cargos ativos por departamento",
        "tags": [
          "Funcionários"
        ]
      }
    },
    "/api/departments": {
      "post": {
        "operationId": "DepartmentsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDepartmentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Departamento criado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Department"
                }
              }
            }
          },
          "400": {
            "description": "Dados de entrada inválidos"
          },
          "409": {
            "description": "Departamento com mesmo nome já existe"
          }
        },
        "summary": "Criar novo departamento",
        "tags": [
          "Departments"
        ]
      },
      "get": {
        "operationId": "DepartmentsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Lista de departamentos retornada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Department"
                  }
                }
              }
            }
          }
        },
        "summary": "Listar todos os departamentos",
        "tags": [
          "Departments"
        ]
      }
    },
    "/api/departments/{id}": {
      "get": {
        "operationId": "DepartmentsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID do departamento",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Departamento encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Department"
                }
              }
            }
          },
          "404": {
            "description": "Departamento não encontrado"
          }
        },
        "summary": "Buscar departamento por ID",
        "tags": [
          "Departments"
        ]
      },
      "patch": {
        "operationId": "DepartmentsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID do departamento",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDepartmentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Departamento atualizado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Department"
                }
              }
            }
          },
          "400": {
            "description": "Dados de entrada inválidos"
          },
          "404": {
            "description": "Departamento não encontrado"
          }
        },
        "summary": "Atualizar departamento por ID",
        "tags": [
          "Departments"
        ]
      },
      "delete": {
        "operationId": "DepartmentsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID do departamento",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Departamento removido com sucesso"
          },
          "404": {
            "description": "Departamento não encontrado"
          },
          "409": {
            "description": "Departamento está sendo usado por funcionários"
          }
        },
        "summary": "Remover departamento por ID",
        "tags": [
          "Departments"
        ]
      }
    },
...
    "/api/positions/department/{departmentId}": {
      "get": {
        "operationId": "PositionsController_findByDepartment",
        "parameters": [
          {
            "name": "departmentId",
            "required": true,
            "in": "path",
            "description": "ID do departamento",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cargos do departamento encontrados",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Position"
                  }
                }
              }
            }
          }
        },
        "summary": "Buscar cargos por departamento",
        "tags": [
          "Positions"
        ]
      }
    },
...
  },
  "info": {
    "title": "Serveon API",
    "description": "API RESTful para o sistema Serveon",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "CreateEmployeeDto": {
        "type": "object",
        "properties": {
          "nome": {
            "type": "string",
            "description": "Nome completo do funcionário",
            "example": "João da Silva"
          },
          "cpf": {
            "type": "string",
            "description": "CPF do funcionário (apenas números)",
            "example": "12345678901"
          },
          "email": {
            "type": "string",
            "description": "Email profissional do funcionário",
            "example": "joao.silva@empresa.com"
          },
          "telefone": {
            "type": "string",
            "description": "Telefone de contato",
            "example": "11987654321"
          },
          "cargoId": {
            "type": "number",
            "description": "ID do cargo do funcionário",
            "example": 1
          },
          "departamentoId": {
            "type": "number",
            "description": "ID do departamento do funcionário",
            "example": 1
          },
          "dataAdmissao": {
            "format": "date-time",
            "type": "string",
            "description": "Data de admissão no formato YYYY-MM-DD",
            "example": "2023-01-15"
          },
          "ativo": {
            "type": "boolean",
            "description": "Status ativo/inativo do funcionário",
            "example": true,
            "default": true
          }
        },
        "required": [
          "nome",
          "cpf",
          "email",
          "dataAdmissao",
          "ativo"
        ]
      },
      "Employee": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "ID único do funcionário",
            "example": 1
          },
          "nome": {
            "type": "string",
            "description": "Nome completo do funcionário",
            "example": "João da Silva"
          },
          "cpf": {
            "type": "string",
            "description": "CPF do funcionário (apenas números)",
            "example": "12345678901"
          },
          "email": {
            "type": "string",
            "description": "Email profissional do funcionário",
            "example": "joao.silva@empresa.com"
          },
          "telefone": {
            "type": "string",
            "description": "Telefone de contato",
            "example": "11987654321"
          },
          "cargoId": {
            "type": "number",
            "description": "ID do cargo do funcionário",
            "example": 1
          },
          "cargoNome": {
            "type": "string",
            "description": "Nome do cargo do funcionário",
            "example": "Vendedor"
          },
          "departamentoId": {
            "type": "number",
            "description": "ID do departamento do funcionário",
            "example": 1
          },
          "departamentoNome": {
            "type": "string",
            "description": "Nome do departamento do funcionário",
            "example": "Comercial"
          },
          "dataAdmissao": {
            "format": "date-time",
            "type": "string",
            "description": "Data de admissão",
            "example": "2023-01-15"
          },
          "dataDemissao": {
            "format": "date-time",
            "type": "string",
            "description": "Data de demissão (se aplicável)",
            "example": null
          },
          "ativo": {
            "type": "boolean",
            "description": "Status ativo/inativo do funcionário",
            "example": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Data de criação do registro",
            "example": "2023-01-15T12:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Data da última atualização do registro",
            "example": "2023-01-15T12:00:00.000Z"
          }
        },
        "required": [
          "id",
          "nome",
          "cpf",
          "email",
          "telefone",
          "cargoId",
          "cargoNome",
          "departamentoId",
          "departamentoNome",
          "dataAdmissao",
          "dataDemissao",
          "ativo",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateEmployeeDto": {
        "type": "object",
        "properties": {
          "dataDemissao": {
            "format": "date-time",
            "type": "string",
            "description": "Data de demissão no formato YYYY-MM-DD",
            "example": "2023-12-31"
          }
        }
      },
      "CreateDepartmentDto": {
        "type": "object",
        "properties": {
          "nome": {
            "type": "string",
            "description": "Nome do departamento",
            "example": "Tecnologia da Informação",
            "maxLength": 100
          },
          "descricao": {
            "type": "string",
            "description": "Descrição do departamento",
            "example": "Departamento responsável pela infraestrutura e desenvolvimento de sistemas",
            "maxLength": 500
          }
        },
        "required": [
          "nome"
        ]
      },
      "Department": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "ID único do departamento",
            "example": 1
          },
          "nome": {
            "type": "string",
            "description": "Nome do departamento",
            "example": "Recursos Humanos",
            "maxLength": 100
          },
          "descricao": {
            "type": "string",
            "description": "Descrição do departamento",
            "example": "Departamento responsável pela gestão de pessoas"
          },
          "ativo": {
            "type": "boolean",
            "description": "Indica se o departamento está ativo",
            "example": true,
            "default": true
          },
          "createdAt": {
            "type": "string",
            "description": "Data de criação do registro",
            "example": "2024-06-08T10:30:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "description": "Data da última atualização do registro",
            "example": "2024-06-08T10:30:00.000Z"
          }
        },
        "required": [
          "id",
          "nome",
          "ativo",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateDepartmentDto": {
        "type": "object",
        "properties": {
          "ativo": {
            "type": "boolean",
            "description": "Status ativo/inativo do departamento",
            "example": true
          }
        }
      },
...
      "Position": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "ID único do cargo",
            "example": 1
          },
          "nome": {
            "type": "string",
            "description": "Nome do cargo",
            "example": "Analista de Sistemas",
            "maxLength": 100
          },
          "descricao": {
            "type": "string",
            "description": "Descrição do cargo",
            "example": "Responsável por desenvolver e manter sistemas"
          },
          "departamentoId": {
            "type": "number",
            "description": "ID do departamento ao qual o cargo pertence",
            "example": 1
          },
          "departamentoNome": {
            "type": "string",
            "description": "Nome do departamento ao qual o cargo pertence",
            "example": "Tecnologia da Informação"
          },
          "ativo": {
            "type": "boolean",
            "description": "Indica se o cargo está ativo",
            "example": true,
            "default": true
          },
          "createdAt": {
            "type": "string",
            "description": "Data de criação do registro",
            "example": "2024-06-08T10:30:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "description": "Data da última atualização do registro",
            "example": "2024-06-08T10:30:00.000Z"
          }
        },
        "required": [
          "id",
          "nome",
          "ativo",
          "createdAt",
          "updatedAt"
        ]
      },
...
    }
  }
}